Conflicts:
	we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/admin/login/service/impl/LoginServiceImpl.java
	we-web-parent/web-linyun-airline/src/main/webapp/WEB-INF/admin/login.jsp
	we-web-parent/web-linyun-airline/src/main/webapp/public/dist/css/AdminLTE.css
	we-web-parent/web-linyun-airline/src/test/java/com/xiaoka/test/GenerateTest.java
	we-web-parent/web-linyun-airline/src/test/resources/code-generator/module.xlsx

+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/admin/authority/function/entity/TFunctionEntity.java
@@ -49,7 +49,7 @@ public class TFunctionEntity implements Serializable {
 
 	@Column
 	@Comment("序号")
-	private long sort;
+	private Integer sort;
 
 	/**在树形节点中是否选中*/
 	private String checked = "false";
